+++ source/renderer/app/themes/overrides/index.js
@@ -1,0 +1,14 @@
+import { IDENTIFIERS } from 'react-polymorph/lib/themes/API';
+import AutocompleteOverrides from './AutocompleteOverrides.scss';
+import ButtonOverrides from './ButtonOverrides.scss';
+import ModalOverrides from './ModalOverrides.scss';
+import SwitchOverrides from './SwitchOverrides.scss';
+
+const { AUTOCOMPLETE, BUTTON, MODAL, SWITCH } = IDENTIFIERS;
+
+export const themeOverrides = {
+  [AUTOCOMPLETE]: AutocompleteOverrides,
+  [BUTTON]: ButtonOverrides,
+  [MODAL]: ModalOverrides,
+  [SWITCH]: SwitchOverrides,
+};